[nvim] adapt colorscheme to follow base16 standard
The previous colorscheme was based on a quite complex one (PaperColor)
and its variable names were not at all consistent with their actual
color contents.
This change uses the base16-oceanicnext [1] colorscheme as a new base to
and adapts the colors to be more in-line with the previously used
PaperColor scheme.
This also greatly improves the highlighting used with the WIP treesitter
plugin.

[1] https://github.com/chriskempson/base16-vim/blob/master/colors/base16-oceanicnext.vim