[office/newsboat] add filter for arxiv feeds
Since newsboat does not handle HTML headers too nicely [1] which is
especially annoying in some scenarios like arXiv feeds (where each
author name is wrapped into an HTML link tag), this makes usage of the
author field in the articlelist-format option almost impossible.

As pointed out by the developer it is possible to use filters [2] in
order to pre-process an XML feed.

Thus, this commit adds a simple Python script which replaces the author
links with their names.

[1]: https://github.com/newsboat/newsboat/issues/403
[2]: https://newsboat.org/releases/2.21/docs/newsboat.html#_scripts_and_filters_snownews_extensions