[tmux] revert to using tmux
Using kitty as a multiplexer is not doable. The only reason why I was
considering it is that I was thinking about using `icat` during some
Jupyter notebook-like development within a kitty pane but I don't think
that is going to really happen.
I even hardly ever display images in the terminal anyways and am
actually going to get rid of all file-manager configs shortly.

I did not see a drawback of using kitty until I noticed that the `ssh`
situation is a pain because kitty's $TERM is not available on most
servers. `kitty +kitten ssh` suggests to fix that, but it does not for
me.

Thus, I'm back with tmux. I am not adding some of the plugins I used
before to try a more bare-bones experience (I don't recall using them
much anyways). I may even consider using sessions more.